requirements.txt
fuse-python
lxml
python-slugify